ContactData and Groupd data moved to model package
