+++ addressbook-web-tests/src/test/java/ru/stqa/pft/addressbook/appmanager/GroupHelper.java
@@ -2,7 +2,7 @@ package ru.stqa.pft.addressbook.appmanager;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
-import ru.stqa.pft.addressbook.test.groups.GroupData;
+import ru.stqa.pft.addressbook.model.GroupData;
 
 public class GroupHelper extends HelperBase {
