+++ addressbook-web-tests/src/test/java/ru/stqa/pft/addressbook/model/ContactData.java
@@ -1,4 +1,4 @@
-package ru.stqa.pft.addressbook.test.contacts;
+package ru.stqa.pft.addressbook.model;
 
 public class ContactData {
     private final String firstName;
