+++ addressbook-web-tests/src/test/java/ru/stqa/pft/addressbook/test/contacts/ContactModificationTests.java
@@ -1,6 +1,7 @@
 package ru.stqa.pft.addressbook.test.contacts;
 
 import org.testng.annotations.Test;
+import ru.stqa.pft.addressbook.model.ContactData;
 import ru.stqa.pft.addressbook.model.TestBase;
 
 public class ContactModificationTests extends TestBase {
