+++ addressbook-web-tests/src/test/java/ru/stqa/pft/addressbook/test/groups/GroupCreationTests.java
@@ -1,6 +1,7 @@
 package ru.stqa.pft.addressbook.test.groups;
 
 import org.testng.annotations.Test;
+import ru.stqa.pft.addressbook.model.GroupData;
 import ru.stqa.pft.addressbook.model.TestBase;
 
 public class GroupCreationTests extends TestBase {
